WIP removes user info modal that is not needed any more bug that a modal is always open
