+++ frontend-nx/apps/edu-hub/components/CourseContent/AchievementRecord.tsx
@@ -6,45 +6,45 @@ import {
   useEffect,
   useState,
 } from 'react';
-import { IUserProfile } from '../../../hooks/user';
-import { useKeycloakUserProfile, useUserId } from '../../../hooks/user';
-import AchievementOptionDropDown from '../../achievements/AchievementOptionDropDown';
+import { IUserProfile } from '../../hooks/user';
+import { useKeycloakUserProfile, useUserId } from '../../hooks/user';
+import AchievementOptionDropDown from '../achievements/AchievementOptionDropDown';
 import {
   AchievementOptionCourses,
   AchievementOptionCoursesVariables,
-} from '../../../queries/__generated__/AchievementOptionCourses';
-import { Button } from '../../common/Button';
-import { useAuthedQuery } from '../../../hooks/authedQuery';
+} from '../../queries/__generated__/AchievementOptionCourses';
+import { Button } from '../common/Button';
+import { useAuthedQuery } from '../../hooks/authedQuery';
 import { BlockTitle } from '@opencampus/shared-components';
-import FormToUploadAchievementRecord from '../../FormToUploadAchievementRecord';
+import FormToUploadAchievementRecord from '../FormToUploadAchievementRecord';
 import {
   makeFullName,
   formattedDate,
   formattedDateWithTime,
-} from '../../../helpers/util';
-import { AlertMessageDialog } from '../../common/dialogs/AlertMessageDialog';
-import { ACHIEVEMENT_OPTION_COURSES } from '../../../queries/achievementOption';
-import { INSERT_AN_ACHIEVEMENT_RECORD } from '../../../queries/achievementRecord';
+} from '../../helpers/util';
+import { AlertMessageDialog } from '../common/dialogs/AlertMessageDialog';
+import { ACHIEVEMENT_OPTION_COURSES } from '../../queries/achievementOption';
+import { INSERT_AN_ACHIEVEMENT_RECORD } from '../../queries/achievementRecord';
 
 import {
   order_by,
   AchievementRecordType_enum,
   AchievementRecordRating_enum,
-} from '../../../__generated__/globalTypes';
+} from '../../__generated__/globalTypes';
 import {
   AchievementRecordListWithAuthors,
   AchievementRecordListWithAuthorsVariables,
   AchievementRecordListWithAuthors_AchievementRecord,
-} from '../../../queries/__generated__/AchievementRecordListWithAuthors';
-import { ACHIEVEMENT_RECORDS_WITH_AUTHORS } from '../../../queries/achievementRecord';
+} from '../../queries/__generated__/AchievementRecordListWithAuthors';
+import { ACHIEVEMENT_RECORDS_WITH_AUTHORS } from '../../queries/achievementRecord';
 import { Link } from '@material-ui/core';
-import { MinAchievementOption } from '../../../helpers/achievement';
+import { MinAchievementOption } from '../../helpers/achievement';
 import useTranslation from 'next-translate/useTranslation';
-import { useAuthedMutation } from '../../../hooks/authedMutation';
+import { useAuthedMutation } from '../../hooks/authedMutation';
 import {
   InsertAnAchievementRecord,
   InsertAnAchievementRecordVariables,
-} from '../../../queries/__generated__/InsertAnAchievementRecord';
+} from '../../queries/__generated__/InsertAnAchievementRecord';
 import Trans from 'next-translate/Trans';
 interface IContext {
   achievementRecordUploadDeadline: any;
